[refactor] API 패키지 변경
